+++ CHANGELOG.md
@@ -44,6 +44,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   tokeniser will now be attempted to be loaded from the base model ID, if available.
 - Now uses EOS token as the PAD token if a generative model has neither PAD nor BOS
   token available.
+- If a generative model has not defined its pad token ID then we now manually check the
+  candidate tokens `<pad>`, `[pad]`, `<|endoftext|>`, `<|im_end|>`, and upper case
+  versions of these tokens.
 
 
 ## [v13.3.0] - 2024-11-29